refactor: rollback removal the currencies source
To verify the generation makes no changes to existing files, this commit
replaces all the original source files from the main branch. After
running the generation script there should be no change.

The `genCurrenciesNumber.mjs` script is also modified to remove the
`@public` comment for api-extractor. This should make the generated
code identical to the originals.